+++ src/main/java/duke/task/Event.java
@@ -1,4 +1,6 @@
-package duke;
+package duke.task;
+
+import duke.utility.DateParser;
 
 import java.time.LocalDateTime;
