+++ src/main/java/duke/utility/DateParser.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.utility;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
