+++ src/main/java/resource/duke.txt
@@ -1,1 +1,2 @@
 D | 1 | Buy groceries | 06-03-2007 13:44
+T | 0 | yo
