+++ src/test/java/duke/ParserTest.java
@@ -1,5 +1,6 @@
 package duke;
 
+import duke.utility.Parser;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
