+++ src/test/java/duke/StorageStub.java
@@ -1,5 +1,8 @@
 package duke;
 
+import duke.task.Task;
+import duke.utility.Storage;
+
 public class StorageStub extends Storage {
     public StorageStub() {
         super("src/main/data/duke.txt");
